% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TelemetryR.r
\docType{data}
\name{ACTtrans}
\alias{ACTtrans}
\title{Transmitters from A.C.T., Updated 2016-01-22}
\format{A data frame with 25 variables and an evolving number of rows
\describe{
  \item{Tag.ID.Code.Standard}{Full transmitter code}
  \item{ID.Standard}{Transmitter number (last digits of transmitter code)}
  \item{Tag.ID.Code.Sensor.I, Tag.ID.Code.Sensor.II}{Optional transmitter
    sensor code}
  \item{ID.Sensor.I, ID.Sensor.II}{Optional transmitter sensor number (last
    digits of sensor code)}
  \item{Release.Date}{Date when tagged individual was released}
  \item{Tag.Life}{Estimated transmitter battery life}
  \item{Expiration.date}{Release date + tag life}
  \item{Tag.Family}{Style/type of tag, or manufaturer if not VEMCO}
  \item{Tag.Type}{Is the tag standard, or does it provide other environmental
    data?}
  \item{Primary.Tagging.Organization}{Organization who attached transmitter}
  \item{Primary.Researcher}{Main point of contact for detections}
  \item{Collaborators}{Auxiliary points of contact for detections}
  \item{Common.Name}{Common name of tagged species}
  \item{Scientific.Name}{Scientific name of tagged species}
  \item{Release.Location}{General location where tagged individual was
    released}
  \item{Pectoral.T.bar, Dorsal.T.bar, PIT.tag, Dart.tag,
    Roto.Flipper.Disc}{Accessory tag identifiers, if used}
  \item{Attachment.method}{How transmitter is attached (external, surgical
    implant, etc.)}
  \item{Comments}{Optional comments by researcher}
}}
\source{
\url{http://www.theactnetwork.com/}
}
\usage{
data(ACTtrans)
}
\description{
Transmitters are submitted to the Atlantic Cooperative Telemetry Network
(\url{http://www.theactnetwork.com/}) by individual researchers and archived
on a private Dropbox folder.
}
\author{
Lori Brown \email{east.coast.telemetry@gmail.com}
}
\keyword{datasets}

