% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injection_time_functions.R
\name{add_injtime}
\alias{add_injtime}
\alias{isolate_postinj}
\title{Add a column for time since injection / break in dataset.}
\usage{
add_injtime(data, injection_time)

isolate_postinj(data, parameter)
}
\arguments{
\item{data}{A dataframe created using the DSI_export_to_dataframe function.}

\item{injection_time}{Time before but close to the injection (24H)}

\item{parameter}{Name of a parameter in the dataset, in quotes.}
}
\value{
A dataframe containing a column for time since injection / break and
columns for each parameter, which each subject synchronized to their
injection time.
}
\description{
This function calculates the time since injection / break in dataset. Use
this function on a dataframe generated from a DSI export using the
DSI_export_to_dataframe function.
}
\section{Functions}{
\itemize{
\item \code{add_injtime}: Add injection time column to dataset.

\item \code{isolate_postinj}: Isolate parameter from dataset after adding injection
time.
}}

