% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injection_time_functions.R
\name{add_injtime}
\alias{add_injtime}
\alias{post_inj_sbp}
\alias{post_inj_dbp}
\alias{post_inj_map}
\alias{post_inj_temp}
\alias{post_inj_HR}
\alias{post_inj_activity}
\title{Add a column for time since injection / break in dataset.}
\usage{
add_injtime(data, injection_time)

post_inj_sbp(data)

post_inj_dbp(data)

post_inj_map(data)

post_inj_temp(data)

post_inj_HR(data)

post_inj_activity(data)
}
\arguments{
\item{data}{A dataframe created using the DSI_export_to_dataframe function.}

\item{injection_time}{Time before but close to the injection (24H)}
}
\value{
A dataframe containing a column for time since injection / break and
columns for each parameter, which each subject synchronized to their
injection time.
}
\description{
This function calculates the time since injection / break in dataset. Use
this function on a dataframe generated from a DSI export using the
DSI_export_to_dataframe function.
}
\section{Functions}{
\itemize{
\item \code{add_injtime}: Add injection time column to dataset.

\item \code{post_inj_sbp}: Isolate SBP from dataset after adding injection time.

\item \code{post_inj_dbp}: Isolate DBP from dataset after adding injection time.

\item \code{post_inj_map}: Isolate MAP from dataset after adding injection time.

\item \code{post_inj_temp}: Isolate temperature from dataset after adding
injection time.

\item \code{post_inj_HR}: Isolate HR from dataset after adding injection time.

\item \code{post_inj_activity}: Isolate activity from dataset after adding injection
time.
}}

\examples{
add_injtime(data = data, lights_on = 21)
post_inj_sbp(data = add_injtime_output)
}
