% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_variable_names}
\alias{api_variable_names}
\title{Names of end-point variables for every API
end-point. Programmatically generated from doc.md}
\format{
An object of class \code{list} of length 35.
}
\usage{
api_variable_names
}
\description{
Names of end-point variables for every API
end-point. Programmatically generated from doc.md
}
\keyword{datasets}
