% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_calculators.R
\name{arb_time_average}
\alias{arb_time_average}
\title{Average telemetry datasets over arbitrary times..}
\usage{
arb_time_average(data, avg_minutes)
}
\arguments{
\item{data}{A dataframe created using the DSI_export_to_dataframe function.}

\item{avg_minutes}{The number of minutes to average over.}
}
\value{
A dataframe containing the numeric columns of the input averaged
over the specified time interval.
}
\description{
This function calculates the average of a parameter over a specified time
period.
}
\section{Functions}{
\itemize{
\item \code{arb_time_average}: Add injection time column to dataset.
}}

\examples{
arb_time_average(sample_BP_data, avg_minutes = 20)
}
