% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBinterp.R
\name{CBinterp}
\alias{CBinterp}
\title{Interpolate variables over the Chesapeake Bay}
\usage{
CBinterp(data, coordinates, res = 2)
}
\arguments{
\item{data}{Data frame or Numeric. Vector of water quality values.}

\item{coordinates}{Numeric. Vector of Long/Lat locations of provided data.}

\item{res}{Numeric. Resolution of the desired interpolated values in
kilometers.}
}
\value{
This function returns an object of class \code{SpatialPointsDataFrame}.
   Data slot contains columns of longitude, latitude, interpolated value at
   that location, and the standard error of the interpolated value.
}
\description{
\code{CBinterp} interpolates Lat-Long georeferenced data across the surface
of the Chesapeake.
}
\details{
This function converts long-lat data to UTM-referenced data, interpolates
across a grid of the required resolution, and returns the interpolated values
as long-lat-referenced data.
}
\examples{
coords <- matrix(c(seq(-77.4, -75.5, length.out = 1000),
                        seq(36.7, 39.4, length.out = 1000)), ncol = 2)
coords <- apply(coords, 2, sample, size = 100)
fake.data <- rnorm(100, 15, 2)
resolution <- 2
interpolated <- CBinterp(data = fake.data, coordinates = coords,
                         res = resolution)
}
\seealso{
\code{\link{chesapeake}}
}

