% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_requests.R, R/payloads.R
\name{list_endpoints}
\alias{list_endpoints}
\alias{list_endpoint_variable}
\alias{check_variables}
\title{Check Endpoint Variables}
\usage{
list_endpoints()

list_endpoint_variable(end_pt = c())

check_variables(x, end_pt)
}
\arguments{
\item{end_pt}{character, the name of the end point}

\item{x}{named vector of variables, where the names are the
variable names, and the values are the value to assign to that
variable}
}
\value{
variable names, endpoint names, or NULL
}
\description{
Convienence functions to list and check the variables for each API
endpoint.
}
\author{
Matt Espe
}
