% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian_avg.R
\name{circadian_avg}
\alias{circadian_avg}
\title{Calculate average parameters during light or dark cycle.}
\usage{
circadian_avg(data, lights_on)
}
\arguments{
\item{data}{A dataframe which is the output from DSI_export_to_dataframe.}

\item{lights_on}{Time when the lights turn on. (24H)}
}
\value{
A list containing two dataframes, one for the averages when lights
are off, and one for averages when lights are on. The first list element
is "Dark Averages" when room lights are OFF, the second list element is
"Light Averages" when room lights are ON.
}
\description{
Theis function takes the output of DSI_export_to_dataframe and calculates
average telemetry parameters during the entire dark or light cycle.
}
\examples{
circadian_avg(data = data, lights_on = 6)
}
