% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_tags.R
\name{comp_tags}
\alias{comp_tags}
\title{Compare tags}
\usage{
comp_tags(v1, v2)
}
\arguments{
\item{v1}{A vector, probably of TagIDs}

\item{v2}{A second vector to compare with the first vector}
}
\value{
Returns a list of two: the first item in the list is the total length of unique things in v1. The second is the total number of unique things in v1 that are also found in v2.
}
\description{
Compare tags
}
\examples{
comp_tags(1:5, 5:10)
comp_tags(v1 = c("A69-1303-63038", "A69-1303-63039", "A69-1303-63040"),
          v2 = c("A69-1303-33085", "A69-1303-33088", "A69-1303-33089"))
          
}
\author{
Myfanwy Johnston, Matt Espe
}
