% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payloads.R
\name{create_antData}
\alias{create_antData}
\title{Create Antenna Data}
\usage{
create_antData(
  df,
  req_cols = c("unATID", "antDataEnd", "unLat", "unLong", "unAntName", "unRegion",
    "antDataStart")
)
}
\arguments{
\item{df}{data.frame}

\item{req_cols}{character vector, the names of the columns required to be in the df}
}
\value{
nested list for conversion into JSON
}
\description{
Creates a nested list from an data.frame to be suitable for
submission to the API
}
\author{
Matt Espe
}
