% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_upload.R
\name{create_registration_data}
\alias{create_registration_data}
\title{Create Registration Data}
\usage{
create_registration_data(
  df,
  data_cols,
  req_cols = c("unAntName", "unTagName", "regDataTime")
)
}
\arguments{
\item{df}{data.frame, with column names exactly "unAntName",
"unTagName", "regDataTime"}

\item{data_cols}{character, the name of columns to use for the
data payload.}

\item{req_cols}{character, the names of columns required to be in
the data.frame}
}
\value{
nested list
}
\description{
A utility for creating a nested list for marshalling to JSON from
a simple data.frame. This function will rarely be called on its
own - instead, it is called from inside \code{api_upload_data()}.
}
\examples{
\dontrun{
create_registration_data(detection_data, data_cols = c("temp_C", "turbidity"))
}
}
\author{
Matt Espe
}
