% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_dist.R
\name{div_dist}
\alias{div_dist}
\title{Divide distance evenly across a units of a time span}
\usage{
div_dist(start, end, distance, time_units = "hour")
}
\arguments{
\item{start}{POSIX, start time}

\item{end}{POSIX, end time}

\item{distance}{numeric, total distance}

\item{time_units}{character, time unit passed to \code{seq.POSIXt}}
}
\value{
data frame with columns \itemize{
  \item date_time: POSIXct column of date/time stamps in the units specified by time_units
  \item prop_dist: the proportion of the distance allocated to each `date_time` row
  }
}
\description{
Divide distance evenly across a units of a time span
}
\examples{
tt = as.POSIXct(c("2020-01-01 00:00:00", "2020-01-02 12:00:00"))
div_dist(tt[1], tt[2], distance = 10)

}
\author{
Myfanwy Johnston
}
