% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download Data}
\usage{
download_data(
  session,
  region,
  speciesID,
  token = session$unToken,
  curl_handle = session$curl,
  api_baseurl = session$base_url,
  db_file = tempfile(),
  end_point = "/api/data/download",
  api_url = paste0(api_baseurl, end_point),
  ...
)
}
\arguments{
\item{session}{list, an object created by \code{start_session()}}

\item{region}{integer, optional??}

\item{speciesID}{integer, optional??}

\item{token}{character, the token for the established session}

\item{curl_handle}{a curl handle for the established session}

\item{api_baseurl}{character, the base URL for the database API}

\item{db_file}{character, the file to store the downloaded data in. Defaults to a temporary file.}

\item{end_point}{character, the suffix for the download end point}

\item{api_url}{character, the full url for the API end point for downloading data}

\item{...}{additional arguments passed to \code{getBinaryURL()}}
}
\value{
character
}
\description{
Downloads the data for a specific user, species, and region into a
temporarly (or not) SQLite3 database. The data can be extracted
from this database via the \code{extract_data()} function.
}
\author{
Matt Espe
}
