% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSI_export_to_dataframe.R
\name{DSI_export_to_dataframe}
\alias{DSI_export_to_dataframe}
\title{Import an Excel file created by DSI's Ponemah export function into R.}
\usage{
DSI_export_to_dataframe(selected_file)
}
\arguments{
\item{selected_file}{Path to an Excel file created by Ponemah.}
}
\value{
A dataframe with 13 columns: .id (SN), Time, TimesOnly, ElapsedTime,
ElapsedTimeMin, ElapsedTimeH, ElapsedTimeD, SBP, DBP, MAP, HR, Temp,
Activity)
}
\description{
This function takes the path to an Excel file as input and generates a
dataframe containing the time, several measures of elapsed time since the
start of the recording, and all telemetry parameters (SBP, DBP, MAP,
HR, Temperature, Activity)
}
\examples{
DSI_export_to_dataframe(selected_file)
}
