% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSI_export_to_dataframe_fast.R
\name{DSI_export_to_dataframe_fast}
\alias{DSI_export_to_dataframe_fast}
\title{Import an Excel file created by DSI's Ponemah export function into R.}
\usage{
DSI_export_to_dataframe_fast(selected_file, output_format = "W")
}
\arguments{
\item{selected_file}{Path to an Excel file created by Ponemah.}

\item{output_format}{W by default. W for wide dataframes, one row per time
point and one column per subject per parameter. L for long dataframes, one
row per time point per subject and one column per parameter.}
}
\value{
A dataframe with 13 columns: .id (SN), Time, TimesOnly, ElapsedTime,
ElapsedTimeMin, ElapsedTimeH, ElapsedTimeD, SBP, DBP, MAP, HR, Temp,
Activity)
}
\description{
This function takes the path to an Excel file as input and generates a
dataframe containing the time, several measures of elapsed time since the
start of the recording, and all telemetry parameters (SBP, DBP, MAP,
HR, Temperature, Activity)
}
