% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{end_session}
\alias{end_session}
\title{End Database Session}
\usage{
end_session(
  session,
  token = session$unToken,
  curl_handle = session$curl,
  api_baseurl,
  end_point = "/api/auth/logout",
  api_url = paste0(api_baseurl, end_point),
  ...
)
}
\arguments{
\item{session}{a list object created via \code{create_session()}}

\item{token}{character, a session token,}

\item{curl_handle}{a curl handle created by
\code{getCurlHandle()}. If one is not provided, it will be
created. Please note, it is best practice to establish a curl
handle when interacting with an API}

\item{api_baseurl}{character, the base URL for the database server}

\item{end_point}{character, the API end point for the login}

\item{api_url}{character, the URL for the database server login}

\item{...}{additional arguments passed to \code{RCurl::getURL()}}

\item{uname}{character}

\item{pwd}{character. See notes for best practices}
}
\description{
Ends the session for an certain token.
}
\author{
Matt Espe
}
