% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract Database Data}
\usage{
extract_data(db_file, env = new.env())
}
\arguments{
\item{db_file}{character, the path to the database file to extract data from.}

\item{env}{an environment to store the tables from the database}
}
\value{
environment, with one data.frame per table in the input database
}
\description{
Extracts all tables from an input database into data.frames inside
an isolated environment. If the \code{env} variable is set to
\code{.GlobalEnv}, the objects will be assigned to variables in
the users global environment. This might be convienent for some,
but has the dangerous side effect of overwritting any existing
objects with the same names in the global environment.
}
\author{
Matt Espe
}
