% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tags.R, R/first_and_last.R
\name{first_and_last}
\alias{first_and_last}
\title{Subset to first (minimum) and last (maximum) detection and stations of a tagged animal}
\usage{
first_and_last(dets_df, tagid_col = "TagID", datetime_col = "DateTimeUTC")

first_and_last(dets_df, tagid_col = "TagID", datetime_col = "DateTimeUTC")
}
\arguments{
\item{dets_df}{A data frame of detections for one or more tags}

\item{tagid_col}{Quoted name of column corresponding to the Tag Identification (TagID, Transmitter, etc)}

\item{datetime_col}{Name of the column (in quotes) corresponding to detection date and time. Default is "DateTimeUTC". Column must be in POSIXct ymd_hm or ymd_hms format.}
}
\value{
a subset of the original dataframe with two rows for each fish (the top row is the minimum detection for that fish, the bottom row is the maximum detection)

a subset of the original dataframe with two rows for each fish (the top row is the minimum detection for that fish, the bottom row is the maximum detection)
}
\description{
Subset to first (minimum) and last (maximum) detection and stations of a tagged animal

Subset to first (minimum) and last (maximum) detection and stations of a tagged animal
}
\details{
in cases where the fish only has one detection associated, or where the first and last detections represent simultaneous detections, only one row will be returned.

in cases where the fish only has one detection associated, or where the first and last detections represent simultaneous detections, only one row will be returned.
}
\author{
Myfanwy Johnston

Myfanwy Johnston
}
