% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_at_final.R
\name{first_at_final}
\alias{first_at_final}
\title{Subset first detection of a fish + first detection at its final known location}
\usage{
first_at_final(
  dets_df,
  tagid_col = "TagID",
  datetime_col = "DateTimeUTC",
  station_col = "Station"
)
}
\arguments{
\item{dets_df}{a dataframe of detections}

\item{tagid_col}{Name of the column (in quotes) corresponding to individual tag IDs}

\item{datetime_col}{Name of the column corresponding to detection date and time (must be in POSIXct ymd_hms format)}

\item{station_col}{Name of the column corresponding to location or station information}
}
\value{
a dataframe with five columns: TagID, first_det (first detection of tag), first_stn (first station),
last_arrival (first detection time at final station), last_stn (station where last arrival was recorded).
}
\description{
Subset first detection of a fish + first detection at its final known location
}
\details{
#' This function takes a dataframe of detections and extracts the first detection of each tag, and the
first detection at each tag's last station, and returns a dataframe with this information so that the travel
time can then be easily calculated
}
\author{
Myfanwy Johnston
}
