% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{GEcircle}
\alias{GEcircle}
\title{Create Google Earth circles around a Lat/Long point}
\usage{
GEcircle(lat, long, radius, color, west = T)
}
\arguments{
\item{lat}{String. Vector of latitudes.}

\item{long}{String. Vector of corresponding longitudes}

\item{radius}{Numeric. Radius, of desired circle in meters.}

\item{color}{Vector of any of the three kinds of R color specifications.
Determines the circle's color; value is passed onto
\code{\link[grDevices]{col2rgb}}.}

\item{west}{Logical. Are the points in the western hemisphere? If values for
\code{long} are already negative, this should be \code{F}.}
}
\value{
This functions will write KML files to your working directory.
}
\description{
\code{GEcircle} creates approximate circles in KML to be opened in Google
Earth
}
\details{
This function uses degree, minute, second format of latitude/longitude points
to allow copy/paste of points from Google Earth. Make sure the degree and
prime symbols are removed and replaced with spaces before running, as in
the example.
}
\examples{
buoys <- data.frame(rbind(c('38 19 1.71', '76 27 4.07'),
                          c('38 18 40.44', '76 18 40.81'),
                          c('38 18 56.44', '76 17 3.80'),
                          c('38 18 27.13', '76 19 55.57')),
                    stringsAsFactors = F)
GEcircle(buoys[,1], buoys[,2], radius = 900, color = 'red', west = T)
}
\seealso{
\code{\link{ptcirc}}
}

