% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_SN_list.R
\name{generate_SN_list}
\alias{generate_SN_list}
\title{Import an Excel file created by DSI's Ponemah export function into R.}
\usage{
generate_SN_list(selected_file)
}
\arguments{
\item{selected_file}{Path to an Excel file created by Ponemah.}
}
\value{
A character vector containing all serial numbers in the Ponemah export.
}
\description{
This function takes the path to an Excel file as input and generates a vector
containing all serial numbers found in the Excel file.
}
\examples{
generate_SN_list(selected_file)
}
