% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_api.R
\name{get_full_endpoint}
\alias{get_full_endpoint}
\title{Get Full Endpoint Name}
\usage{
get_full_endpoint(
  endpoint_shortname,
  endpoint_fullnames = names(api_variable_names)
)
}
\arguments{
\item{endpoint_shortname}{character, the short name to match}

\item{endpoint_fullnames}{character vector, the full names of the
endpoints to match to}
}
\value{
character, the matched full name of the endpoint
}
\description{
A convenience function for returning the full endpoint name from a
partial (short) name.
}
\author{
Matt Espe
}
