% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_full_dataset.R
\name{isolate_parameter}
\alias{isolate_parameter}
\title{Isolate parameters from complete telemetry dataframe.}
\usage{
isolate_parameter(data, parameter, mean_col = FALSE)
}
\arguments{
\item{data}{A dataframe which is the output from DSI_export_to_dataframe.}

\item{parameter}{Name of a parameter to isolate in the dataframe.}

\item{mean_col}{A boolean for including a new column with the group mean for
the selected parameter.}
}
\value{
A dataframe with the Time and elapsed_time columns from the input,
and a column for each subject appended with the parameter (i.e. SBP).
Additionally an optional column for the mean of all subjects in the dataset.
}
\description{
These functions take the output of DSI_export_to_dataframe and outputs a new
dataframe for the specified parameter.
}
\section{Functions}{
\itemize{
\item \code{isolate_parameter}: Isolate parameter from complete telemetry
dataframe export.
}}

\examples{
isolate_parameter(sample_BP_data, SBP, mean_col = TRUE)
isolate_parameter(sample_BP_data, MAP)
}
