% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_full_dataset.R
\name{isolate_sbp}
\alias{isolate_sbp}
\alias{isolate_dbp}
\alias{isolate_map}
\alias{isolate_temp}
\alias{isolate_HR}
\alias{isolate_activity}
\title{Isolate parameters from complete telemetry dataframe.}
\usage{
isolate_sbp(data)

isolate_dbp(data)

isolate_map(data)

isolate_temp(data)

isolate_HR(data)

isolate_activity(data)
}
\arguments{
\item{data}{A dataframe which is the output from DSI_export_to_dataframe.}
}
\value{
A dataframe with the Time and elapsed_time columns from the input,
and a column for each subject appended with the parameter (i.e. SBP).
Additionally a column for the mean of all subjects in the dataset.
}
\description{
These functions take the output of DSI_export_to_dataframe and outputs a new
dataframe for the specified parameter.
}
\section{Functions}{
\itemize{
\item \code{isolate_sbp}: Isolate SBP from complete dataframe.

\item \code{isolate_dbp}: Isolate DBP from complete dataframe.

\item \code{isolate_map}: Isolate MAP from complete dataframe.

\item \code{isolate_temp}: Isolate temperature from complete dataframe.

\item \code{isolate_HR}: Isolate HR from complete dataframe.

\item \code{isolate_activity}: Isolate activity from complete dataframe.
}}

\examples{
isolate_sbp(exported_data)
isolate_dbp(exported_data)
isolate_map(exported_data)
isolate_temp(exported_data)
isolate_HR(exported_data)
isolate_activity(exported_data)
}
