% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_typical_day.R
\name{isolate_typical}
\alias{isolate_typical}
\title{Isolate parameters from typical_day dataframe.}
\usage{
isolate_typical(data, parameter)
}
\arguments{
\item{data}{A dataframe which is the output from typical_day.}

\item{parameter}{Name of a parameter in the dataset, in quotes.}
}
\value{
A dataframe with the Time columns from the input,
and a column for each subject appended with the parameter (i.e. SBP).
 @examples
isolate_typical(data = typical_day_output, parameter = "SBP")
isolate_typical(data = typical_day_output, parameter = "HR")
}
\description{
Theis function takes the output of typical_day and outputs a new
dataframe for the specified parameter.
}
