% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_inj_min.R
\name{post_inj_min}
\alias{post_inj_min}
\title{Calculate the minimum value of parameters after injection.}
\usage{
post_inj_min(data, duration)
}
\arguments{
\item{data}{A dataframe created using the DSI_export_to_dataframe function.}

\item{duration}{Maximum time after injection to look for minimum (hours).}
}
\value{
A dataframe containing the minimum value  of the parameter for each
SN.
}
\description{
This function calculates the minimum value of telemetry parameters after
an injection. The input for this function is the output of the post_inj
functions, i.e. post_inj_sbp.
}
