% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptcirc.R
\name{ptcirc}
\alias{ptcirc}
\title{Draw circles around a Lat/Long point}
\usage{
ptcirc(lonlatpoint, radius)
}
\arguments{
\item{lonlatpoint}{Numeric vector. A set of longitude and latitude pairs
(decimal degrees), in that order. Can be a data frame or matrix, but the
first column must be the longitude and the second must be latitude.}

\item{radius}{Numeric. Radius of desired circle, in m.}
}
\value{
Output is a data frame with longitude, latitude, and "circle"
   columns. The circle column demarcates circles belonging to different
   Lon/Lat pairs.
}
\description{
\code{ptcirc} creates a data frame of points forming a circle around a point.
}
\details{
The function was adopted from an answer on R-Help, 20061107, by Arien Lam
\url{https://stat.ethz.ch/pipermail/r-help/2006-November/116851.html}.
Each point is the place where you end up if you travel a certain distance
along a great circle, which is uniquely defined by a point (your
starting point) and an angle with the meridian at that point (your
direction).
}
\examples{
ptcirc(c(-75, 37), 1609)
ptcirc(data.frame(lon = c(-76, -80), lat = c(34, 37)), 2000)
}
\seealso{
\code{\link{GEcircle}}
}

