% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo_an.R
\name{quo_an}
\alias{quo_an}
\title{Habitat selectivity through quotient analysis}
\usage{
quo_an(wq, det, bin_width = 1, pres_abs = F, R = 999)
}
\arguments{
\item{wq}{Numeric. Water quality data at each station.}

\item{det}{Numeric. Number of detections at each station.}

\item{bin_width}{Numeric. Size of water quality bins. Default is 1.}

\item{pres_abs}{Logical. Should the data be reduced to presence/absence?
Default is false.}

\item{R}{Numeric. Number of bootstrap replicates. Default is 999.}
}
\value{
Output is a data frame with bin labels, number of detections or
               number sites with detections (depending on \code{pres_abs}
               input) within each bin, number of sites within each bin, and
               pMe, pSe, and Qe as defined above. Confidence intervals at the
               0.025 and 0.975 percentiles are also provided.
}
\description{
\code{quo_an} executes the quotient approach defined in Overholtz et
al. 2011. Proportion of fish incidence in a water quality bin is divided by
the proportion of stations within that bin. Values greater than 1 indicate
preference.
}
\details{
This approach was outlined in:

Van Der Lingen, CD, L Hutchings, D Merkle, JJ van der Westhuizen, and J
Nelson. 2001. Comparative spawning habitats of anchovy (\emph{Engraulis
capensis}) and sardine (\emph{Sardinops sagax}) in the southern Benguela
upwelling ecosystem. 185-209. Spatial processes and management of marine
populations. Alaska Sea Grant College Program, AK-SG-01-02, Fairbanks.

It was defined, as it appears here, in:

Overholtz, W.J., Hare, J.A. and Keith, C.M., 2011. Impacts of interannual
environmental forcing and climate change on the distribution of Atlantic
mackerel on the US Northeast continental shelf. Marine and Coastal
Fisheries, 3(1), pp.219-232.

The proportion of sites in each water quality bin (\emph{pSe}) is
determined by dividing the number of stations within the bin by the total
number of stations. The porportion of stations with fish (\emph{M}) in
each bin (\emph{pMe}) is determined by dividing the number of stations in the
bin with fish by the total number of stations with fish. The quotient value
for the water quality bin (\emph{Qe}) is calculated as \emph{pMe/pSe}.

Values greater than 1 \dQuote{indicate a greater number of positive (fish)
stations than expected based on sampling effort}. Confidence intervals are
calculated using \code{boot.ci} from the \code{boot} package. The \code{type}
input to \code{boot.ci} is "\code{perc}"; all other inputs to \code{boot} and
\code{boot.ci} are the function defaults.
}

