% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_BP_data.R
\docType{data}
\name{sample_BP_data}
\alias{sample_BP_data}
\title{Example telemetry data from one cohort of mice implanted with blood
pressure telemeters.}
\format{A data frame with 42495 rows and 13 variables:
\describe{
  \item{Time}{Time stamp of observation, POSIXct}
  \item{TimesOnly}{Time of day from Time column, ITime}
  \item{ElapsedTime}{Time elapsed since recording began, seconds}
  \item{1024071_SBP}{Systolic blood pressure for SN 1024071, mmHg}
  \item{1024071_DBP}{Diastolic blood pressure for SN 1024071, mmHg}
  \item{1024071_MAP}{Mean arterial blood pressure for SN 1024071, mmHg}
  \item{1024071_HR}{Heart rate for SN 1024071, bpm}
  \item{1024071_Temp}{Temperaturefor SN 1024071, degrees C}
  \item{1024071_Activity}{Activity score for SN 1024071, arbitrary units}
  \item{1024084_SBP}{Systolic blood pressure for SN 1024084, mmHg}
  \item{1024084_DBP}{Diastolic blood pressure for SN 1024084, mmHg}
  \item{1024084_MAP}{Mean arterial blood pressure for SN 1024084, mmHg}
  \item{1024084_HR}{Heart rate for SN 1024084, bpm}
  \item{1024084_Temp}{Temperaturefor SN 1024084, degrees C}
  \item{1024084_Activity}{Activity score for SN 1024084, arbitrary units}
  \item{1033468_SBP}{Systolic blood pressure for SN 1033468, mmHg}
  \item{1033468_DBP}{Diastolic blood pressure for SN 1033468, mmHg}
  \item{1033468_MAP}{Mean arterial blood pressure for SN 1033468, mmHg}
  \item{1033468_HR}{Heart rate for SN 1033468, bpm}
  \item{1033468_Temp}{Temperaturefor SN 1033468, degrees C}
  \item{1033468_Activity}{Activity score for SN 1033468, arbitrary units}

}}
\description{
A dataset containing time and blood pressure telemeter parameters for 15 mice
over approximately 4 days.
}
\author{
Andrew Stiegler \email{andrew.stiegler@gmail.com}
}
