% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{start_session}
\alias{start_session}
\title{Start Database Session}
\usage{
start_session(
  uname,
  pwd,
  api_baseurl,
  end_point = "/api/auth/login",
  api_url = paste0(api_baseurl, end_point),
  curl_handle,
  ...
)
}
\arguments{
\item{uname}{character}

\item{pwd}{character. See notes for best practices}

\item{api_baseurl}{character, the base URL for the database server}

\item{end_point}{character, the API end point for the login}

\item{api_url}{character, the URL for the database server login}

\item{curl_handle}{a curl handle created by
\code{getCurlHandle()}. If one is not provided, it will be
created. Please note, it is best practice to establish a curl
handle when interacting with an API}

\item{...}{additional arguments passed to \code{RCurl::getURL()}}
}
\value{
list
}
\description{
Login to the database server using pre-established login
credentials. This will return a token for the session, which is
used in all subsequent calls the the API.
}
\details{
Best practices are to avoid storing your password in Rscripts that
might be shared. To avoid doing this, you can set the credentials
in a separate file, and read that file into R to set the login
credentials.
}
\author{
Matt Espe
}
