% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sturgrow.R
\name{sturgrow}
\alias{sturgrow}
\title{Bioenergetics model for juvenile Atlantic sturgeon}
\usage{
sturgrow(TEMP, SAL, DO, GR = 14, TL = NULL, CJ_OBS = 0, RATION = 1,
  t1 = 6, t4 = 28, tk1rm = 0.141, tk4rm = 0.796, b1 = -0.158,
  hrm = 0.268, irm = 0.352, smin = 9.166, do1 = 25, grm = 0.748,
  crm = 1, drm = 1.048, arm = 0.522, brm = -0.17, ox = 13.55,
  tl98fc = 26.09, tk1fc = 0.195, tk4fc = 0.556, jfc = 0.359,
  kfc = 0.247, s1 = 1, s4 = 29, gfc = 0.733, dfc = 2.516, cfc = 1,
  afc = 1.028, bfc = -0.197, aeg = 0.335, ceg = -0.75, deg = -0.62,
  geg = 0, aex = 0.055703, bex = -0.29, cex = 0.0392, asda = 0.1657,
  aact = 0.29)
}
\arguments{
\item{TEMP}{Numeric vector. Recorded temperature (celsius).}

\item{SAL}{Numeric vector. Corresponding recorded salinity.}

\item{DO}{Numeric vector. Corresponding recorded dissolved oxygen
(percent saturation)}

\item{GR}{Numeric. Weight of fish (kilograms). Defaults to 14 kg.}

\item{TL}{Numeric. Total length of fish (cm). Defaults to NULL.}

\item{CJ_OBS}{Numeric. If you declare this variable equal to 0, the model
assumes maximum consumption rate (CMAX). Defaults to 0.}

\item{RATION}{Numeric. If you declare this variable equal to 1, the model
assumes p-vaule equal to 1. Defaults to 1.}

\item{t1}{Numeric. Lowest tested temperature (celsius). Constant in the
temperature portion of the routine metabolism and food consumption
models. Defaults to 6.}

\item{t4}{Numeric. Highest tested temperature (celsius). Constant in the
temperature portion of the routine metabolism and food consumption
models. Defaults to 28.}

\item{tk1rm}{Numeric. Reaction rate multiplier at the lowest tested
temperature (6 C). Parameter of temperature portion of routine metabolism
model. Defaults to 0.141.}

\item{tk4rm}{Numeric. Lower temperature threshold where f_rm(TEMP) >= 0.98.
Parameter of temperature portion of routine metabolism model. Defaults to
0.796.}

\item{b1}{Numeric. Specific gill surface area. Constant in salinity portion
  of routine metabolism and food consumption models. Defaults to -0.158
  after work of Burggren et al (1979) on \strong{Acipenser transmontanus}.
@param hrm Numeric. Hyperosmotic response coefficient. Parameter of 
  salinity portion of routine metabolism model. Defaults to 0.268.}

\item{irm}{Numeric. Hyposmotic response coefficient. Parameter of
salinity portion of routine metabolism model. Defaults to 0.352.}

\item{smin}{Numeric. Salinity at which minimum osmoregulation cost is
predicted. Parameter of salinity portion of routine metabolism and food
consumption models. Defaults to 9.166.}

\item{do1}{Numeric. Constant in the dissolved oxygen portion of the routine
metabolism and food consumption models. ***UPDATE***}

\item{grm}{Numeric. Proportionality constant for DOC_rm. Parameter of
dissolved oxygen portion of routine metabolism model. Defaults to 0.748.}

\item{crm}{Numeric. Dissolved oxygen response shape parameter. Parameter of
dissolved oxygen portion of routine metabolism model. Defaults to 1.0.}

\item{drm}{Numeric. Proportionality constant for reaction rate at lowest
DO_sat. Parameter of dissolved oxygen portion of routine metabolism
model. Defaults to 1.048.}

\item{arm}{Numeric. Allometric intercept (scaling coefficient), parameter of
routine metabolism model. Defaults to 0.522.}

\item{brm}{Numeric. Allometric slope, parameter of routine metabolism model.
Defaults to -0.17.}

\item{ox}{Numeric. Constant in routine metabolism model ***UPDATE***}

\item{tl98fc}{Numeric. Lower temperature threshold where f_fc(T) >= 0.98.
Parameter of temperature portion of food consumption model. Defaults to
26.09.}

\item{tk1fc}{Numeric. Reaction rate multiplier at the lowest tested
temperature (6 C). Parameter of temperature portion of food consumption
model. Defaults to 0.195.}

\item{tk4fc}{Numeric. Reaction rate multiplier at the highest tested
temperature (28 C). Parameter of temperature portion of food consumption
model. Defaults to 0.556.}

\item{jfc}{Numeric. Size-dependent intercept for reaction rate at the lowest
salinity. Parameter of salinity portion of food consumption model.
Defaults to 0.359.}

\item{kfc}{Numeric. Size-dependent intercept for reaction rate at the
highest salinity. Parameter of salinity portion of food consumption
model. Defaults to 0.247.}

\item{s1}{Numeric. Lowest tested salinity. Constant in the salintiy portion
of the food consumption model. Defaults to 1.}

\item{s4}{Numeric. Highest tested salinity. Constant in the salinity portion
of the food consumption model. Defaults to 29.}

\item{gfc}{Numeric. Proportionality constant for DOC_fc. Parameter of
dissolved oxygen portion of food consumption model. Defaults to 0.733.}

\item{dfc}{Numeric. Proportionality constant for reaction rate at lowest
DO_sat. Parameter of dissolved oxygen portion of food consumption model.
Defaults to 2.516.}

\item{cfc}{Numeric. Dissolved oxygen response shape parameter. Parameter of
dissolved oxygen portion of food consumption model. Defaults to 1.0.}

\item{afc}{Numeric. Allometric intercept (scaling coefficient), parameter of
food consumption model. Defaults to 1.028.}

\item{bfc}{Numeric. Allometric slope, parameter of food consumption model.
Defaults to -0.197.}

\item{aeg}{Numeric. Scale parameter for egestion. Parameter of egestion
model. Defaults to 0.335.}

\item{ceg}{Numeric. Dissolved oxygen effect exponent. Parameter of egestion
model. Defaults to -0.75.}

\item{deg}{Numeric. Temperature effect exponent. Parameter of egestion
model. Defaults to -0.62.}

\item{geg}{Numeric. Ration size effect exponent. Parameter of egestion
model. Defaults to 0.}

\item{aex}{Numeric. RNE, scaling factor. Parameter of excretion model.
Defaults to 0.055703.}

\item{bex}{Numeric. RNE, exponent. Parameter of excretion model. Defaults to
-0.29.}

\item{cex}{Numeric. XNE, food consumption proportionality coefficient.
Parameter of excretion model. Defaults to 0.0392.}

\item{asda}{Numeric. Proportionality constant (to assimilated energy).
Constant of post-prandial metabolism (SDA) model. Defaults to 0.1657.}

\item{aact}{Numeric. Proportionality constant to food consumption. Constant
of active metabolism (ACT) model. Defaults to 0.29.}
}
\value{
Output is the predicted growth in the respective temperature,
   salintiy, and dissolved oxygen combinations.
}
\description{
\code{sturgrow} executes the bioenergetics model from Niklitschek &
Secor 2009.
}
\details{
For more information on model development and full explanation of
parameters, see the supporting material of Niklitschek E.J., Secor D.H.,
2009. Dissolved oxygen, temperature and salinity effects on the
ecophysiology and survival of juvenile Atlantic sturgeon in estuarine
waters: II. Model development and testing. Journal of Experimental Marine
Biology and Ecology 381: S161-S172.

Script was originally written on 20110803, last update on 20140702.
Manipulated into R fuction on 20150330 by M. O'Brien 
\email{obrien@umces.edu}.
}
\author{
Edwin Niklitschek \email{edwin.niklitschek@ulagos.cl}
}

