\name{telemetry-package}
\alias{telemetry-package}
\alias{telemetry}
\docType{package}
\title{
\packageTitle{telemetry}
}
\description{
\packageDescription{telemetry}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{telemetry}
\packageIndices{telemetry}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{telemetry}

Maintainer: \packageMaintainer{telemetry}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
