% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetRy.R
\docType{package}
\name{telemetRy}
\alias{telemetRy}
\title{telemetRy: A package for interacting with telemetry data in R.}
\description{
This is a package for interacting with telemetry data using R. Specific
functions for importing data from DSI telemetry systems are included, as well
as functions for analyzing telemetry data over various timescales.
}
\section{Functions}{

DSI_export_to_dataframe
A simple function for bringing exports from DSI's Ponemah software into R.
Takes an Excel file as its input, and returns a dataframe.

typical_day
This function takes a dataframe with a time column as input. The function
generates a "typical day" dataframe, with a time column with the temporal
resolution of the input, and calculates the average of multiple
observations at each timepoint.

circadian_avg
This function takes a dataframe with multiple circadian cycles as input, and
outputs a dataframe averaging each column over the dark or light circadian
cycle.
}

