% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Create coarse movement tracks.}
\usage{
track(data, dates, ids)
}
\arguments{
\item{data}{Data frame. Location of date/time and identifier values.}

\item{dates}{character. Column name of date/time values within \code{data}.}

\item{ids}{character. Column name(s) of identifiers within \code{data}.
Multiple names may be provided in a vector: c('station', 'lat', 'long'),
e.g.}
}
\value{
Output is a data frame with date/times at unique combinations of
   provided IDs.
}
\description{
\code{tracks} returns locations (station, array, etc.) in the order in which
they were visited.
}
\details{
This function orders detection data by date/time and only returns detections
where the fish moved from one identified location to another.
}
