% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typical_day.R
\name{typical_day}
\alias{typical_day}
\alias{typical_average}
\title{Calculate average parameters for each timepoint in a circadian cycle.}
\usage{
typical_day(data, lights_on)

typical_average(data, avg_minutes)
}
\arguments{
\item{data}{A dataframe created using the DSI_export_to_dataframe function.}

\item{lights_on}{Time when the lights turn on. (24H)}

\item{avg_mintes}{Number of minutes to average over (integer)}
}
\value{
A dataframe containing a column of circadian time and parameters.
}
\description{
This function calculates the typical parameters in a dataframe for one
circadian cycle. Use this function on a dataframe generated from a DSI export
 using the DSI_export_to_dataframe function.
}
\section{Functions}{
\itemize{
\item \code{typical_day}: Calculate typical day from telemetry data.

\item \code{typical_average}: Average typical_day over specified time interval.
}}

\examples{
typical_day(data = data, lights_on = 21)
typical_day(data = data, lights_on = 6)
typical_average(data = data, avg_minutes = 20)
}
