% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_typical_day.R
\name{typical_sbp}
\alias{typical_sbp}
\alias{typical_dbp}
\alias{typical_map}
\alias{typical_temp}
\alias{typical_HR}
\alias{typical_activity}
\title{Isolate parameters from typical_day dataframe.}
\usage{
typical_sbp(data)

typical_dbp(data)

typical_map(data)

typical_temp(data)

typical_HR(data)

typical_activity(data)
}
\arguments{
\item{data}{A dataframe which is the output from typical_day.}
}
\value{
A dataframe with the Time and elapsed_time columns from the input,
and a column for each subject appended with the parameter (i.e. SBP).
Additionally a column for the mean of all subjects in the dataset.
}
\description{
These functions take the output of typical_day and outputs a new
dataframe for the specified parameter.
}
\section{Functions}{
\itemize{
\item \code{typical_sbp}: Isolate SBP from typical_day dataframe.

\item \code{typical_dbp}: Isolate DBP from typical_day dataframe.

\item \code{typical_map}: Isolate MAP from typical_day dataframe.

\item \code{typical_temp}: Isolate temperature from typical_day dataframe.

\item \code{typical_HR}: Isolate HR from typical_day dataframe.

\item \code{typical_activity}: Isolate activity from typical_day dataframe.
}}

\examples{
typical_sbp(typical_day_output)
typical_dbp(typical_day_output)
typical_map(typical_day_output)
typical_temp(typical_day_output)
typical_HR(typical_day_output)
typical_activity(typical_day_output)
}
