% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vemsort.R
\name{vemsort}
\alias{vemsort}
\title{Prepare VEMCO transmitter CSV files for analysis}
\usage{
vemsort(directory = getwd(), clust = NULL, prog_bar = F,
  creation_date = NULL)
}
\arguments{
\item{directory}{String. Location of CSV data, defaults to current wd.}

\item{clust}{A cluster object created by \code{\link[parallel]{makeCluster}}.
If cluster is supplied, this will use \code{\link[parallel]{parLapply}} to
import the files. Defaults to NULL with no parallel evaluation.}

\item{prog_bar}{Logical. Do you want a progress bar displayed? Will increase
evaluation time. Initiates \code{\link[pbapply]{pblapply}}.}

\item{creation_date}{Character date in a standard unambiguous format
(e.g., YYYY-MM-DD). Will select only files created after this date.}
}
\value{
Output is a data frame containing all detections from
   the directory's CSV files. Adds two columns: one containing local time of
   the detections (as defined by \code{Sys.timzone}) and one containing the
   detection's CSV file of origin.
}
\description{
\code{vemsort} finds and combines all VEMCO CSV files in a directory
}
\details{
This function assumes that all necessary CSV files are within the specified
directory or subdirectories within. All files must have the default headings
offloaded by VEMCO products. These are, in order:
Date and Time (UTC), Receiver, Transmitter, Transmitter Name,
Transmitter Serial, Sensor Value, Sensor Unit, Station Name,
Latitude, Longitude.
}
\examples{
vemsort('C:/Users/mypcname/Documents/Vemco/Vue/ReceiverLogs')

# Select files created after Jan 1, 2015
vemsort('C:/Users/mypcname/Documents/Vemco/Vue/ReceiverLogs',
         creation_date = '2015-01-01')

# Use parallel computation and a progress bar
cl <- parallel::makeCluster(parallel::detectCores() - 1)
vemsort('C:/Users/mypcname/Documents/Vemco/Vue/ReceiverLogs',
         clust = cl, prog_bar = T)
parallel::stopCluster(cl)
}
\seealso{
\code{\link[parallel]{makeCluster}}, \code{\link[parallel]{parLapply}},
   \code{\link[pbapply]{pblapply}}
}
