% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACTsplit.R
\name{ACTsplit}
\alias{ACTsplit}
\title{Split detections into CSV files according to ACT database}
\usage{
ACTsplit(directory = getwd(), ACTtrans, my.trans = NULL, write = TRUE,
  out = NULL, start = 20040101, end = Sys.Date())
}
\arguments{
\item{directory}{String passed on to \code{\link{vemsort}}. Location of CSV
data, which defaults to current working directory.}

\item{ACTtrans}{character. Location of ACT transmitter database.}

\item{my.trans}{Numeric vector. Tag ID codes that you want removed prior
to distribution.}

\item{write}{Logical. Do you want to output CSV files? Useful if you are
only looking for unidentified detections.}

\item{out}{String. Where do you want the CSV files placed?}

\item{start}{Numeric. Date in ymd form from which you want to start pulling
detections, assumed to be Eastern time zone. Defaults to Jan 1, 2000.}

\item{end}{Numeric. Date in ymd form to stop pulling detections, assumed to
be Eastern time zone. Defaults to current system date.}
}
\value{
Outputs are CSV files in the form of ResearcherCurrentdate.csv and a
   data frame containing detections of unidentified codes. The CSV files
   will be found in your current working directory by default.
}
\description{
\code{ACTsplit} filters all detections according to the researchers in the
\href{http://www.theactnetwork.com/}{ACT Network}
}
\details{
This function uses \code{\link{vemsort}} to import detections, then matches
these detections with transmitters submitted to ACT. It also allows a
specified date range and removal of personal or private transmitters.
}
\examples{
ACTsplit('C:/Users/mypcname/Documents/Vemco/Vue/ReceiverLogs')
# Choose detections from April 1, 2014 up to and including August 1, 2014.
ACTsplit('C:/Users/MYPCNAME/Documents/Vemco/Vue/ReceiverLogs',
           start = 20140401, end = 20140801)
}
\seealso{
\code{\link{vemsort}}, \code{\link{ACTupdate}}
}
