% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACTupdate.R
\name{ACTupdate}
\alias{ACTupdate}
\title{Download ACT data base.}
\usage{
ACTupdate(keep = F)
}
\arguments{
\item{keep}{logical. Should you keep the downloaded xlsx? Default is \code{F}.}
}
\value{
Saves a copy of the current "Active transmitters mm-dd-yy.xlsx"
   spreadsheet to your working directory in RData format as "ACTtrans.rda".
   If keep == T, it also will save the orignial .xlsx.
}
\description{
\code{ACTupdate} connects to the
\href{http://www.theactnetwork.com/}{ACT Network} Dropbox data base,
downloads the transmitter list, and converts the list to RData format. In
order to access Dropbox, you'll need to log in and create an authorization
token. The token can be saved for later use.
}
\details{
The function checks if a local version of the ACT transmitter data base
exists. If it does not, or if the local version is older than the Dropbox
version, it downloads the file from Dropbox. The .xlsx is converted to RData
and deleted unless keep == T.
}
\examples{
ACTupdate()
}
\seealso{
\code{\link{vemsort}}, \code{\link{ACTsplit}}
}

