% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_api.R
\name{check_variables}
\alias{check_variables}
\alias{list_endpoints}
\alias{list_endpoint_variable}
\title{Check Endpoint Variables}
\usage{
check_variables(x, end_pt)

list_endpoints()

list_endpoint_variable(end_pt = c())
}
\arguments{
\item{x}{named vector of variables, where the names are the
variable names, and the values are the value to assign to that
variable}

\item{end_pt}{character, the name of the end point}
}
\value{
variable names, endpoint names, or NULL
}
\description{
Convienence functions to list and check the variables for each API
endpoint.
}
\author{
Matt Espe
}
