% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payloads.R
\name{create_antenna_data}
\alias{create_antenna_data}
\alias{create_batch_tag_data}
\title{Create Antenna Data}
\usage{
create_antenna_data(
  df,
  req_cols = c("unATID", "antDataEnd", "unLat", "unLong", "unAntName", "unRegion",
    "antDataStart")
)

create_batch_tag_data(
  df,
  payload_cols,
  req_cols = c("unTagName", "tagDataTime")
)
}
\arguments{
\item{df}{data.frame}

\item{req_cols}{character vector, the names of the columns required to be in the df}

\item{payload_cols}{character vector, the column names of data to
include in the data payload. These should match the values
provided to the batchDataSchema fields.}
}
\value{
nested list for conversion into JSON
}
\description{
Creates a nested list from an data.frame of antennas to be
suitable for submission to the API endpoint
/api/admin/create/network.
}
\author{
Matt Espe
}
