% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payloads.R
\name{create_payload}
\alias{create_payload}
\title{Create API Data Payload}
\usage{
create_payload(end_pt, ...)
}
\arguments{
\item{end_pt}{character, the name of the API end point to create a payload for}

\item{...}{key = value pairs for creating the JSON payload}
}
\value{
list
}
\description{
This function creates a nested list in the correct structure to be
converted to JSON and delivered to the API
}
\author{
Matt Espe
}
