% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_dist.R
\name{div_dist}
\alias{div_dist}
\title{Divide distance evenly across a units of a time span}
\usage{
div_dist(start, end, distance, time_units = "hour")
}
\arguments{
\item{start}{}

\item{end}{}

\item{distance}{}

\item{time_units}{}
}
\value{
data frame with columns \itemize{
  \item date_time: POSIXct column of date/time stamps in the units specified by time_units
  \item prop_dist: the proportion of the distance allocated to each `date_time` row
  }
}
\description{
Divide distance evenly across a units of a time span
}
