% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download Data}
\usage{
download_data(
  session,
  region,
  speciesID,
  token = session$unToken,
  curl_handle = session$curl,
  api_baseurl = session$base_url,
  db_file = tempfile(),
  end_point = "/api/data/download",
  api_url = paste0(api_baseurl, end_point),
  ...,
  debug_json = FALSE
)
}
\arguments{
\item{session}{list, an object created by \code{start_session()}}

\item{region}{integer, optional??}

\item{speciesID}{integer, optional??}

\item{token}{character, the token for the established session}

\item{curl_handle}{a curl handle for the established session}

\item{api_baseurl}{character, the base URL for the database API}

\item{db_file}{character, the file to store the downloaded data
in. Defaults to a temporary file.}

\item{end_point}{character, the suffix for the download end point}

\item{api_url}{character, the full url for the API end point for
downloading data}

\item{...}{additional arguments passed to \code{getBinaryURL()}}

\item{debug_json}{logial, if TRUE the function returns the JSON
payload as a string and does not send the request. This is
primarily for debugging API behavior and checking the JSON
payloads for correctness. The JSON will include white space for
pretty printing, e.g. \code{cat(payload)}}
}
\value{
character, the file path to the downloaded SQLite
  database, or when \code{debug_json = TRUE} a string representing
  the JSON payload.
}
\description{
Downloads the data for a specific user, species, and region into a
temporarly (or not) SQLite3 database. The data can be extracted
from this database via the \code{extract_data()} function.
}
\examples{
\dontrun{
# All data
detection_db = download_data(my_session)

# One species
detection_db = dowloand(data(my_session, speciesID = 1)

# defaults to putting tables in an environment
detection_data = extract_data(detection_db)

# Access data via $
names(detection_data)
detection_data$Registrations
as.list(detection_data)
}
}
\author{
Matt Espe
}
