% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orphans.R
\name{find_orphans}
\alias{find_orphans}
\title{Identify orphaned detections between deployment and detection records}
\usage{
find_orphans(
  detections_df,
  deployments_df,
  deployment_start_col = "DeploymentStart",
  deployment_end_col = "DeploymentEnd",
  deployment_rec_col = "Receiver",
  detections_df_datetime_col = "DateTimePST",
  fuzzy_match = 60 * 60,
  ...
)
}
\arguments{
\item{detections_df}{a dataframe of detections}

\item{deployments_df}{a data frame of receiver deployments}

\item{fuzzy_match}{desired time threshold to expand the deployment windows by in order to accommodate slight mismatches between deployment windows and detections}

\item{...}{arguments passed on to data.table::foverlaps()}
}
\value{
data frame of detections with deployment data brought in, where NAs in the ExpandedStart and ExpandedEnd columns indicate orphaned detections.
}
\description{
Identify orphaned detections between deployment and detection records
}
\details{
relies on data.table::foverlaps() to accomplish the merge.  Warnings will issue if you have detections with no matching deployments, or if you have detections with multiple matching deployments (some of which are caused by the fuzzy_match).
}
\author{
Myfanwy Johnston
}
