% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_requests.R
\name{send_api_request}
\alias{send_api_request}
\title{Send Database API request}
\usage{
send_api_request(
  session,
  ...,
  end_point,
  token = session$unToken,
  curl_handle = session$curl,
  api_baseurl = session$base_url,
  api_url = paste0(api_baseurl, end_point)
)
}
\arguments{
\item{session}{list, and active session object created by \code{start_session()}}

\item{...}{key = value pairs of parameters for each end point
request. These vary by endpoint. For a list of the required
variables for each end point, use
\code{list_endpoint_variable()}}

\item{end_point}{chracter, the end point for the API}

\item{token}{character, the unToken for an active session}

\item{curl_handle}{an existing curl handle for an active
session. If not provided, one will be created just for this
call.}

\item{api_baseurl}{character, the base URL for the API}

\item{api_url}{character, the full URL to the endpoint}
}
\value{
json result
}
\description{
This is a flexible function for interacting with any of the
numerous /api/* endpoints. It functions by taking key
= value pairs required for each of the endpoints and crafting the
appropriate JSON payload to submit to the server.
}
\details{
There are some particularities around how the database expects
some elements to be created. See the vignette
"Telemetry API Administration" for details.
}
\author{
Matt Espe
}
