% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typical_day.R
\name{typical_day}
\alias{typical_day}
\title{Calculate average parameters for each timepoint in a circadian cycle.}
\usage{
typical_day(data, lights_on)
}
\arguments{
\item{data}{A dataframe created using the DSI_export_to_dataframe function..}

\item{lights_on}{Time when the lights turn on. (24H, as integer)}
}
\value{
A dataframe containing a column of circadian time and parameters.
}
\description{
This function calculates the typical parameters in a dataframe for one
circadian cycle. Use this function on a dataframe generated from a DSI export
 using the DSI_export_to_dataframe function.
}
\examples{
typical_day(data = data, lights_on = 21)
typical_day(data = data, lights_on = 6)
}
