% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{vemsort}
\alias{vemsort}
\title{Prepare VEMCO transmitter CSV files for analysis}
\usage{
vemsort(directory = getwd(), false.det = NULL)
}
\arguments{
\item{directory}{String. Location of CSV data, defaults to current wd.}

\item{false.det}{Numeric vector. Contains tag ID codes of known
false detections.}
}
\value{
Output is a dplyr table data frame containing all detections from
   the directory's CSV files
}
\description{
\code{vemsort} finds and combines all VEMCO CSV files in a directory
}
\details{
This function assumes that all necessary CSV files are within the specified
directory or subdirectories within. All files must have the default headings
offloaded by VEMCO products. These are, in order:
Date and Time (UTC), Receiver, Transmitter, Transmitter Name,
Transmitter Serial, Sensor Value, Sensor Unit, Station Name,
Latitude, Longitude.
}
\examples{
vemsort('C:/Users/mypcname/Documents/Vemco/Vue/ReceiverLogs')
vemsort('C:/Users/mypcname/Documents/Vemco/Vue/ReceiverLogs',
         c('37119', '64288'))
}

